safer member deletion

+++ app/controllers/admin/members_controller.rb
@@ -67,9 +67,11 @@ class Admin::MembersController < ApplicationController
     assocs_destroyed = 0
     params[:emails_for_deletion].each_line do |line|
       if member = Member.find_by_dirty_email_address(line)
-        assocs_destroyed += destroy_associations(member)
-        member.destroy
-        members_destroyed += 1
+        unless member.good_contact?
+          assocs_destroyed += destroy_associations(member)
+          member.destroy
+          members_destroyed += 1
+        end
       end
     end
     redirect_to admin_members_url, notice: "#{members_destroyed} Member(s) and #{assocs_destroyed} associated records (event_mail_logs, member_tokens, rsvps) were successfully destroyed."
